Move image API builtin implementation
