+++ src/api/llm/builtins/imageApi.js
